+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 -Changes
 	- update header and footer default font size
 	- update and improve `utils.request` function [breaking changes][WIP]
+- Fixes
+	- `utils.request` now return proper text error message if possible
 
 2.7.0 - 2025-10-23
 - Feat: